Refactor restapi package structure
