--- src/main/java/org/synyx/urlaubsverwaltung/restapi/DepartmentsListWrapper.java
@@ -1,22 +1,0 @@
-package org.synyx.urlaubsverwaltung.restapi;
-
-import java.util.List;
-
-
-/**
- * @author  David Schilling - schilling@synyx.de
- */
-class DepartmentsListWrapper {
-
-    private final List<DepartmentResponse> departments;
-
-    DepartmentsListWrapper(List<DepartmentResponse> departments) {
-
-        this.departments = departments;
-    }
-
-    public List<DepartmentResponse> getDepartments() {
-
-        return departments;
-    }
-}
