--- src/main/java/org/synyx/urlaubsverwaltung/restapi/WorkDayResponse.java
@@ -1,19 +1,0 @@
-package org.synyx.urlaubsverwaltung.restapi;
-
-/**
- * @author  David Schilling - schilling@synyx.de
- */
-class WorkDayResponse {
-
-    private final String workDays;
-
-    WorkDayResponse(String workDays) {
-
-        this.workDays = workDays;
-    }
-
-    public String getWorkDays() {
-
-        return workDays;
-    }
-}
