+++ src/main/java/org/synyx/urlaubsverwaltung/restapi/workday/WorkDayResponse.java
@@ -1,0 +1,19 @@
+package org.synyx.urlaubsverwaltung.restapi.workday;
+
+/**
+ * @author  David Schilling - schilling@synyx.de
+ */
+class WorkDayResponse {
+
+    private final String workDays;
+
+    WorkDayResponse(String workDays) {
+
+        this.workDays = workDays;
+    }
+
+    public String getWorkDays() {
+
+        return workDays;
+    }
+}
